requirements.txt
streamlit
openai
numpy
Pillow
